add jarak in order detail response

+++ formatters/mitraFormat.go
@@ -23,7 +23,7 @@ type KatalogResponse struct{
 	Name string `json:"name"`
 	MinPrice float64 `json:"minimal_price"`
 	MaxPrice float64 `json:"maximal_price"`
-	Distance int `json:"jarak"`
+	Distance float64 `json:"jarak"`
 	Bidang []BidangResponse `json:"bidang"`
 	Rating int `json:"rating"`
 }
